+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [Unreleased] - ReleaseDate
+### Changed
+- `Mutex::with`, `RwLock::with_read`, and `RwLock::with_write` now require
+  their Futures to be `Send`.  They also can return errors.
+  ([#7](https://github.com/asomers/futures-locks/pull/7))
+
+### Fixed
+- `Mutex::with`, `RwLock::with_read`, and `RwLock::with_write` now work with
+  all Tokio Runtimes.
+  ([#7](https://github.com/asomers/futures-locks/pull/7))
+
 ## [0.2.1] - 2018-06-18
 ### Changed
 - Tokio support is now enabled by default.